# docs/swagger.yaml
definitions:
  api.Message:
    properties:
      code:
        type: integer
      message:
        type: string
    type: object
  api.Meta:
    properties:
      filtered_count:
        type: integer
      total_count:
        type: integer
    type: object
  api.SwgErrResponse:
    properties:
      error:
        type: string
      messages:
        items:
          $ref: '#/definitions/api.Message'
        type: array
    type: object
...
  api.SwgResponse-array_bookview_BaseView:
    properties:
      data:
        items:
          $ref: '#/definitions/bookview.BaseView'
        type: array
      messages:
        items:
          $ref: '#/definitions/api.Message'
        type: array
      meta:
        $ref: '#/definitions/api.Meta'
    type: object
...
  api.SwgResponse-array_userview_SimpleView:
    properties:
      data:
        items:
          $ref: '#/definitions/userview.SimpleView'
        type: array
      messages:
        items:
          $ref: '#/definitions/api.Message'
        type: array
      meta:
        $ref: '#/definitions/api.Meta'
    type: object
...
  api.SwgResponse-bookview_BaseView:
    properties:
      data:
        $ref: '#/definitions/bookview.BaseView'
      messages:
        items:
          $ref: '#/definitions/api.Message'
        type: array
      meta:
        $ref: '#/definitions/api.Meta'
    type: object
  api.SwgResponse-bookview_DetailedView:
    properties:
      data:
        $ref: '#/definitions/bookview.DetailedView'
      messages:
        items:
          $ref: '#/definitions/api.Message'
        type: array
      meta:
        $ref: '#/definitions/api.Meta'
    type: object
...
  api.SwgResponse-loanview_DetailedView:
    properties:
      data:
        $ref: '#/definitions/loanview.DetailedView'
      messages:
        items:
          $ref: '#/definitions/api.Message'
        type: array
      meta:
        $ref: '#/definitions/api.Meta'
    type: object
  api.SwgResponse-reservationview_DetailedView:
    properties:
      data:
        $ref: '#/definitions/reservationview.DetailedView'
      messages:
        items:
          $ref: '#/definitions/api.Message'
        type: array
      meta:
        $ref: '#/definitions/api.Meta'
    type: object
...
  bookparams.CreateParams:
    properties:
      author:
        type: string
      genre:
        type: string
      isbn:
        type: string
      language:
        type: string
      publication_date:
        type: string
      publisher:
        type: string
      title:
        type: string
    type: object
  bookparams.UpdateParams:
    properties:
      author:
        type: string
      genre:
        type: string
      id:
        type: integer
      isbn:
        type: string
      language:
        type: string
      publication_date:
        type: string
      publisher:
        type: string
      title:
        type: string
    type: object
  bookview.BaseView:
    properties:
      author:
        type: string
      genre:
        type: string
      id:
        type: integer
      isbn:
        type: string
      language:
        type: string
      publication_date:
        type: string
      publisher:
        type: string
      title:
        type: string
    type: object
  bookview.DetailedView:
    properties:
      author:
        type: string
      genre:
        type: string
      id:
        type: integer
      isbn:
        type: string
      language:
        type: string
      loans:
        items:
          $ref: '#/definitions/loanview.BaseView'
        type: array
      publication_date:
        type: string
      publisher:
        type: string
      reservations:
        items:
          $ref: '#/definitions/reservationview.BaseView'
        type: array
      title:
        type: string
    type: object
  fineview.View:
    properties:
      amount:
        type: number
      id:
        type: integer
      loan_id:
        type: integer
      status:
        type: string
      user_id:
        type: integer
    type: object
  loanhistoryview.View:
    properties:
      action:
        type: string
      id:
        type: integer
      loan_id:
        type: integer
    type: object
  loanview.BaseView:
    properties:
      book_id:
        type: integer
      borrow_date:
        type: string
      due_date:
        type: string
      fines:
        items:
          $ref: '#/definitions/fineview.View'
        type: array
      id:
        type: integer
      loan_histories:
        items:
          $ref: '#/definitions/loanhistoryview.View'
        type: array
      return_date:
        type: string
      status:
        type: string
      user_id:
        type: integer
    type: object
  loanview.BookView:
    properties:
      author:
        type: string
      genre:
        type: string
      id:
        type: integer
      isbn:
        type: string
      language:
        type: string
      publication_date:
        type: string
      publisher:
        type: string
      title:
        type: string
    type: object
  loanview.DetailedView:
    properties:
      book:
        $ref: '#/definitions/loanview.BookView'
      book_id:
        type: integer
      borrow_date:
        type: string
      due_date:
        type: string
      fines:
        items:
          $ref: '#/definitions/fineview.View'
        type: array
      id:
        type: integer
      loan_histories:
        items:
          $ref: '#/definitions/loanhistoryview.View'
        type: array
      return_date:
        type: string
      status:
        type: string
      user:
        $ref: '#/definitions/userview.View'
      user_id:
        type: integer
    type: object
...
  personview.View:
    properties:
      full_name:
        type: string
      id:
        type: integer
      language_preference:
        type: string
      preferred_name:
        type: string
    type: object
  reservationview.BaseView:
    properties:
      book_id:
        type: integer
      id:
        type: integer
      reservation_date:
        type: string
      status:
        type: string
      user_id:
        type: integer
    type: object
  reservationview.BookView:
    properties:
      author:
        type: string
      genre:
        type: string
      id:
        type: integer
      isbn:
        type: string
      language:
        type: string
      publication_date:
        type: string
      publisher:
        type: string
      title:
        type: string
    type: object
  reservationview.DetailedView:
    properties:
      book:
        $ref: '#/definitions/reservationview.BookView'
      book_id:
        type: integer
      id:
        type: integer
      reservation_date:
        type: string
      status:
        type: string
      user:
        $ref: '#/definitions/userview.View'
      user_id:
        type: integer
    type: object
...
  userview.SimpleView:
    properties:
      id:
        type: integer
      username:
        type: string
    type: object
  userview.View:
    properties:
      abilities:
        items:
          type: string
        type: array
      email:
        type: string
      id:
        type: integer
      person_attributes:
        $ref: '#/definitions/personview.View'
      username:
        type: string
    type: object
info:
  contact: {}
  title: Library Mangement System API
paths:
  /api/v1//loan/{loan_id}/:
    get:
      consumes:
      - '*/*'
      description: Retrieves a loan from the library
      parameters:
      - description: loan ID to read
        in: path
        name: loan_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.SwgResponse-loanview_DetailedView'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.SwgErrResponse'
      summary: Read a loan
      tags:
      - loan
...
  /api/v1/book:
    get:
      consumes:
      - application/json
      description: Lists books in the library
      parameters:
      - description: Offset for pagination
        in: query
        name: offset
        type: integer
      - description: Limit for pagination
        in: query
        name: limit
        type: integer
      - description: Filter by value
        in: query
        name: filter[value]
        type: string
      - description: Sort by column name (e.g. title)
        in: query
        name: sortBy
        type: string
      - description: Order by direction (asc or desc)
        in: query
        name: orderBy
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.SwgResponse-array_bookview_BaseView'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.SwgErrResponse'
      summary: List books
      tags:
      - book
  /api/v1/book/:
    post:
      consumes:
      - application/json
      description: creates a new book in the library
      parameters:
      - description: Book creation request
        in: body
        name: book
        required: true
        schema:
          $ref: '#/definitions/bookparams.CreateParams'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.SwgResponse-bookview_BaseView'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.SwgErrResponse'
      summary: Create a book
      tags:
      - book
  /api/v1/book/{book_id}:
    delete:
      consumes:
      - '*/*'
      description: deletes a book from the library
      parameters:
      - description: Book ID to delete
        in: path
        name: book_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.SwgResponse-bookview_BaseView'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.SwgErrResponse'
      summary: Delete a book
      tags:
      - book
    get:
      consumes:
      - '*/*'
      description: reads a book in the library
      parameters:
      - description: Book ID to read
        in: path
        name: book_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.SwgResponse-bookview_DetailedView'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.SwgErrResponse'
      summary: Read book
      tags:
      - book
    patch:
      consumes:
      - application/json
      description: Updates an existing book in the library
      parameters:
      - description: Book ID to update
        in: path
        name: book_id
        required: true
        type: integer
      - description: Book update request
        in: body
        name: book
        required: true
        schema:
          $ref: '#/definitions/bookparams.UpdateParams'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.SwgResponse-bookview_BaseView'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.SwgErrResponse'
      summary: Update a book
      tags:
      - book
  /api/v1/book/{book_id}/loan/:
    post:
      consumes:
      - '*/*'
      description: Loans a book from the library
      parameters:
      - description: Book ID for loan
        in: path
        name: book_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.SwgResponse-loanview_DetailedView'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.SwgErrResponse'
      summary: Loan a book
      tags:
      - loan
...
  /api/v1/book/autocomplete:
    get:
      consumes:
      - '*/*'
      description: Autocomplete 10 books based on input value
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.SwgResponse-array_userview_SimpleView'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.SwgErrResponse'
      summary: Autocomplete 10 books based on input value
      tags:
      - user
...
  /api/v1/loan/{loan_id}/:
    delete:
      consumes:
      - '*/*'
      description: Deletes an existing loan in the library
      parameters:
      - description: loan ID to delete
        in: path
        name: loan_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.SwgResponse-loanview_DetailedView'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.SwgErrResponse'
      summary: Delete a loan
      tags:
      - loan
...
  /api/v1/loan/{loan_id}/return:
    patch:
      consumes:
      - '*/*'
      description: Returns a book to the library
      parameters:
      - description: loan ID to return
        in: path
        name: loan_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.SwgResponse-loanview_DetailedView'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.SwgErrResponse'
      summary: Return a book
      tags:
      - loan
...
  /api/v1/reservation/{reservation_id}/cancel:
    patch:
      consumes:
      - '*/*'
      description: Cancels a reservation for a book
      parameters:
      - description: reservation ID to cancel
        in: path
        name: reservation_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.SwgResponse-reservationview_DetailedView'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.SwgErrResponse'
      summary: Cancel a reservation
      tags:
      - reservation
  /api/v1/reservation/{reservation_id}/checkout:
    patch:
      consumes:
      - '*/*'
      description: Checks out a book for a given reservation
      parameters:
      - description: reservation ID to checkout
        in: path
        name: reservation_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.SwgResponse-reservationview_DetailedView'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.SwgErrResponse'
      summary: Checkout a book
      tags:
      - reservation
...
swagger: "2.0"